refactor: move llm webhook http implementations internal
